+++ baseweb_components/streamlit_baseweb/__init__.py
@@ -13,10 +13,15 @@ if not _RELEASE:
     _base_web_modal = components.declare_component(
         "base_web_modal", url="http://localhost:3000",
     )
+    _base_web_button = components.declare_component(
+        "base_web_button", url="http://localhost:3001",
+    )
 else:
     parent_dir = os.path.dirname(os.path.abspath(__file__))
     modal_dir = os.path.join(parent_dir, "modal/build")
     _base_web_modal = components.declare_component("base_web_modal", path=modal_dir)
+    button_dir = os.path.join(parent_dir, "button/build")
+    _base_web_button = components.declare_component("base_web_button", path=button_dir)
 
 
 def base_web_modal(
@@ -27,19 +32,23 @@ def base_web_modal(
     size: Literal["full", "default", "auto"] = "full",
     validation_button_label: str = "Okay",
     key=None,
-):
+) -> None:
     """
+    Defines a modal component in the Base Web framework.
 
-    Parameters
-    ----------
-    title: modal's title
-    body: modal's body
-    is_open: modal's initial state
-    role: options are 'dialog' or 'alertdialog'
-    size: options are 'default', 'full', 'auto'
-    validation_button_label: label to display for validation button
+    Parameters:
     -------
+    - title (str): The title of the modal.
+    - body (str): The body content of the modal.
+    - is_open (bool, optional): Whether the modal is open or not. Default is True.
+    - role (Literal["dialog", "alertdialog"], optional): The ARIA role of the modal. Default is "dialog".
+    - size (Literal["full", "default", "auto"], optional): The size of the modal. Default is "full".
+    - validation_button_label (str, optional): The label of the validation button. Default is "Okay".
+    - key (any, optional): A key to identify the modal. Default is None.
 
+    Returns:
+    -------
+    None
     """
     _base_web_modal(
         title=title,
@@ -81,3 +90,53 @@ def base_web_modal(
     }    
         """
     st.sidebar.markdown(f"<style>{modal_css}</style>", unsafe_allow_html=True)
+
+
+def base_web_button(
+    label: str = "Submit",
+    disabled: bool = False,
+    size: Literal["default", "compact", "large"] = "default",
+    shape: Literal["default", "pill", "round", "square"] = "default",
+    kind: Literal["primary", "secondary", "tertiary"] = "primary",
+    key=None,
+) -> bool:
+    """
+    Defines a button component in the Base Web framework.
+
+    Parameters:
+    -------
+    - label : str, optional
+        The label or text to display on the button. Default is "Submit".
+    - disabled : bool, optional
+        Whether the button should be disabled or not. Default is False.
+    - size : str, optional
+        The size of the button. Available options are "default", "compact", and "large".
+        Default is "default".
+    - shape : str, optional
+        The shape of the button. Available options are "default", "pill", "round", and "square".
+        Default is "default".
+    - kind : str, optional
+        The kind or style of the button. Available options are "primary", "secondary", "tertiary",
+        "minimal", and "outline". Default is "primary".
+    - key : Any, optional
+        A unique identifier for the button. Default is None.
+
+    Returns:
+    ----------
+    True or False
+
+    """
+    # TODO: to be integrated in React:
+    #  - colors overrides in React
+    #  - startEnhancer
+    #  - endEnhancer
+    if _base_web_button(
+        label=label,
+        disabled=disabled,
+        size=size,
+        shape=shape,
+        kind=kind,
+        key=key,
+        default=0,
+    ):
+        return True